add ready valid to to mux const

+++ tests/common/test_mux_aoi.py
@@ -50,7 +50,8 @@ def test_aoi_mux_wrapper(height, width):
 @pytest.mark.parametrize('height,width', [(choice([3, 5, 6, 7, 9]),
                                            randint(1, 32))
                                           for _ in range(5)])
-def test_aoi_const_mux_wrapper(height, width):
+@pytest.mark.parametrize("mux_type", [AOIMuxType.Const, AOIMuxType.ConstReadyValid])
+def test_aoi_const_mux_wrapper(height, width, mux_type):
     """
     Test that the mux wrapper circuit works as expected. Specifically, we
     initialize a mux with random height and width, and check that the output is
@@ -59,27 +60,35 @@ def test_aoi_const_mux_wrapper(height, width):
     Note that we do not check the behavior with sel > height, because this is
     undefined behavior.
     """
-    mux = AOIMuxWrapper(height, width, AOIMuxType.Const)
+    mux = AOIMuxWrapper(height, width, mux_type)
     assert mux.width == width
     assert mux.name() == \
-        f"MuxWrapperAOI_{height}_{width}_{AOIMuxType.Const.name}"
+        f"MuxWrapperAOI_{height}_{width}_{mux_type.name}"
 
     mux_circuit = mux.circuit()
     tester = fault.Tester(mux_circuit)
     inputs = [fault.random.random_bv(width) for _ in range(height + 1)]
+    valid_inputs = [fault.random.random_bv(1) for _ in range(height + 1)]
     for i, input_ in enumerate(inputs[:height]):
         tester.poke(mux_circuit.I[i], input_)
+    if mux_type == AOIMuxType.ConstReadyValid:
+        for i, vin in enumerate(valid_inputs[:height]):
+            tester.poke(mux_circuit.valid_in[i], vin)
     for i in range(height + 1):
         tester.poke(mux_circuit.S, BitVector[mux.sel_bits](i))
         tester.eval()
         if i < height:
             tester.expect(mux_circuit.O, inputs[i])
+            if mux_type == AOIMuxType.ConstReadyValid:
+                tester.expect(mux_circuit.valid_out, valid_inputs[i])
         else:
             tester.expect(mux_circuit.O, 0)
 
     with tempfile.TemporaryDirectory() as tempdir:
         for aoi_mux in glob.glob("tests/common/rtl/*.sv"):
-            shutil.copy(aoi_mux, tempdir)
+            if not os.path.exists(os.path.join(tempdir, os.path.basename(aoi_mux))):
+                shutil.copy(aoi_mux, tempdir)
         tester.compile_and_run(directory=tempdir,
                                magma_output="coreir-verilog",
                                flags=["-Wno-fatal"])
+
